PIM-5094 Rewied Invalid item registry
