+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/archiving.yml
@@ -8,13 +8,13 @@ parameters:
     pim_base_connector.archiver.archivable_file_writer_archiver.class: Pim\Bundle\BaseConnectorBundle\Archiver\ArchivableFileWriterArchiver
     pim_base_connector.archiver.invalid_item_csv_writer.class: Pim\Bundle\BaseConnectorBundle\Archiver\CsvInvalidItemWriter
     pim_base_connector.archiver.invalid_item_xlsx_writer.class: Pim\Bundle\BaseConnectorBundle\Archiver\XlsxInvalidItemWriter
-    pim_base_connector.event_listener.invalid_items_writer_register.class: Pim\Bundle\BaseConnectorBundle\EventListener\InvalidItemWriterResolver
+    pim_base_connector.event_listener.invalid_items_writer_registry.class: Pim\Bundle\BaseConnectorBundle\EventListener\InvalidItemWriterRegistry
 
 services:
     pim_base_connector.event_listener.archivist:
         class: '%pim_base_connector.event_listener.archivist.class%'
         arguments:
-              - '@pim_base_connector.event_listener.invalid_items_writer_register'
+              - '@pim_base_connector.event_listener.invalid_items_writer_registry'
         tags:
             - { name: kernel.event_subscriber }
 
@@ -80,5 +80,5 @@ services:
         tags:
             - { name: pim_base_connector.invalid_items_writer }
 
-    pim_base_connector.event_listener.invalid_items_writer_register:
-        class: '%pim_base_connector.event_listener.invalid_items_writer_register.class%'
+    pim_base_connector.event_listener.invalid_items_writer_registry:
+        class: '%pim_base_connector.event_listener.invalid_items_writer_registry.class%'
